Update pal from commit: 8ae1695e
* Add a 'BitIter' helper
* Add multikeyboard support to PAL GPU Profiler on Linux
* Apply trimming to CmdAllocator
* Make freeing Deque<PodType> faster
* App-detect initial DCC clear value
* Implement VK_KHR_maintenance4
* palHashLiteralString: fix bytes >= 0x80
* Add UserData SpillTable support for ExecuteIndirect packet
* Add a ClockMode setting to the GPU Profiler
* New Util::Span
* Fix bug in ValidateGraphicsUserData
* Fix coord warnings when using SDMA upload path w/ Framebenches
*  Copy metadata header by pfp to solve RenderDoc corruption
* Redirect Decorator's GetTraceSession()
* Re-write spill table if pipeline changed
* Debug logging support in PAL
* Fix gles cts failure 'GLES3.functional.polygon_offset.fixed16_render_with_units', vulkan as angle backend

+++ inc/util/palHashLiteralString.h
@@ -56,7 +56,8 @@ public:
     static constexpr uint32 Hash(
         const char (&string)[N])
     {
-        return LowPart(static_cast<uint64>(Fnv1aHashHelper<N, I - 1>::Hash(string) ^ string[I - 1]) * Fnv1aPrime);
+        return
+            LowPart(static_cast<uint64>(Fnv1aHashHelper<N, I - 1>::Hash(string) ^ uint8(string[I - 1])) * Fnv1aPrime);
     }
 };
 
@@ -74,7 +75,7 @@ public:
     static constexpr uint32 Hash(
         const char (&string)[N])
     {
-        return LowPart(static_cast<uint64>(Fnv1aOffset ^ string[0]) * Fnv1aPrime);
+        return LowPart(static_cast<uint64>(Fnv1aOffset ^ uint8(string[0])) * Fnv1aPrime);
     }
 };
